Login edited, content added to pages
