--- src/modules/Login/LoginSeparator/index.js
@@ -1,11 +1,0 @@
-import React from 'react'
- 
-import styles from './style.module.css'
- 
-const LoginSeparator = () => (
-    <section className={styles.separator}>
-        <h2>LOGIN</h2>
-    </section>
-)
- 
-export default LoginSeparator
