--- src/modules/Login/LoginSeparator/style.module.css
@@ -1,16 +1,0 @@
-.separator {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100px;
-    background-color: var(--main-color);
-}
-
-
-    .separator h2 {
-        font-size: 36px;
-        padding: 0 20px;
-        color: white;
-        text-transform: uppercase;
-        font-weight: lighter;
-    }
