+++ src/pages/login.module.css
@@ -81,4 +81,3 @@
     .success.show {
         opacity: 1;
     }
-
